v0018
Game auto-detection has been added which is performed when user logs in.
This functionality currently works only on distributions of Windows
operating system.
This functionality requires on server side that in gameslist.dat file is
set path and key (second one is omitted if value is set with
default/null key) in registry that contains path of directory (or path
to file) where game executable is located.
Also a problem has occurred with interpreting of mentioned file because
probably methods for manipulation on binary data has changed in some of
later versions of Qt framework so currently is used slower workaround.