v0031
-when performing game auto-detection on Windows, no longer are generated
empty unnecessary keys when actual requested keys are not found (current
solution possibly causes performance-loss on that operation, but at
least it doesn't leave garbage in registry)
-partial game auto-detection is now also supported on Linux (supported
are only Linux games (i.e. not those that are run via Wine) and for each
supported game has to be defined default directory where the game's
executable file is possibly located, e.g. /usr/games/, ~/X/Y, ...)
-null-strings are removed from all string-literals as they are
automatically appended
-if application detects that the game, which user is playing,
communicates with destination with address 0.0.0.0 (or with some other
address inside 0.0.0.0/24 range which all can't be used for addressing
devices by RFC3330), that address is not shared as other users could not
join user over it
-messages in program for maintaining list of supported games are now
more concise as for game auto-detection purposes is now on Linux used
default path of executable file and on Windows is used full path of the
key that contains game's installation directory as a value